Added barrelling for current time
